Revert "Revert "Merge pull request #890 from GitGuardian/garancegourdel/scrt-4513-ggshield-investigate-run-out-of-memory""
This reverts commit 23f453619fff7638c0af7ba668002613ea4335d6.
